requirements.txt
# requirements.txt
pdf2image
pandas
openai
Pillow
pytesseract